# gen/httpapi/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package httpapi

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/v1/images/"

			if l := len("/v1/images/"); len(elem) >= l && elem[0:l] == "/v1/images/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "confirm"
				origElem := elem
				if l := len("confirm"); len(elem) >= l && elem[0:l] == "confirm" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "POST":
						s.handleConfirmUploadRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "POST")
					}

					return
				}

				elem = origElem
			case 'p': // Prefix: "pr"
				origElem := elem
				if l := len("pr"); len(elem) >= l && elem[0:l] == "pr" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'e': // Prefix: "esign"

					if l := len("esign"); len(elem) >= l && elem[0:l] == "esign" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleCreatePresignRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}

				case 'o': // Prefix: "omote"

					if l := len("omote"); len(elem) >= l && elem[0:l] == "omote" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handlePromoteImagesRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}

				}

				elem = origElem
			}
			// Param: "id"
			// Match until "/"
			idx := strings.IndexByte(elem, '/')
			if idx < 0 {
				idx = len(elem)
			}
			args[0] = elem[:idx]
			elem = elem[idx:]

			if len(elem) == 0 {
				switch r.Method {
				case "DELETE":
					s.handleDeleteImageRequest([1]string{
						args[0],
					}, elemIsEscaped, w, r)
				case "GET":
					s.handleGetImageRequest([1]string{
						args[0],
					}, elemIsEscaped, w, r)
				default:
					s.notAllowed(w, r, "DELETE,GET")
				}

				return
			}
			switch elem[0] {
			case '/': // Prefix: "/url"

				if l := len("/url"); len(elem) >= l && elem[0:l] == "/url" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleGetDeliveryUrlRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/v1/images/"

			if l := len("/v1/images/"); len(elem) >= l && elem[0:l] == "/v1/images/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "confirm"
				origElem := elem
				if l := len("confirm"); len(elem) >= l && elem[0:l] == "confirm" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "POST":
						r.name = ConfirmUploadOperation
						r.summary = "Confirm upload (HEAD against S3), persist metadata"
						r.operationID = "confirmUpload"
						r.operationGroup = ""
						r.pathPattern = "/v1/images/confirm"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}

				elem = origElem
			case 'p': // Prefix: "pr"
				origElem := elem
				if l := len("pr"); len(elem) >= l && elem[0:l] == "pr" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'e': // Prefix: "esign"

					if l := len("esign"); len(elem) >= l && elem[0:l] == "esign" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = CreatePresignOperation
							r.summary = "Create a POST policy for uploading an image to MinIO/S3 with size validation"
							r.operationID = "createPresign"
							r.operationGroup = ""
							r.pathPattern = "/v1/images/presign"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'o': // Prefix: "omote"

					if l := len("omote"); len(elem) >= l && elem[0:l] == "omote" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = PromoteImagesOperation
							r.summary = "Promote images from a product draft to a final product"
							r.operationID = "promoteImages"
							r.operationGroup = ""
							r.pathPattern = "/v1/images/promote"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

				elem = origElem
			}
			// Param: "id"
			// Match until "/"
			idx := strings.IndexByte(elem, '/')
			if idx < 0 {
				idx = len(elem)
			}
			args[0] = elem[:idx]
			elem = elem[idx:]

			if len(elem) == 0 {
				switch method {
				case "DELETE":
					r.name = DeleteImageOperation
					r.summary = "Delete image (soft or hard, per policy)"
					r.operationID = "deleteImage"
					r.operationGroup = ""
					r.pathPattern = "/v1/images/{id}"
					r.args = args
					r.count = 1
					return r, true
				case "GET":
					r.name = GetImageOperation
					r.summary = "Get image metadata"
					r.operationID = "getImage"
					r.operationGroup = ""
					r.pathPattern = "/v1/images/{id}"
					r.args = args
					r.count = 1
					return r, true
				default:
					return
				}
			}
			switch elem[0] {
			case '/': // Prefix: "/url"

				if l := len("/url"); len(elem) >= l && elem[0:l] == "/url" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch method {
					case "GET":
						r.name = GetDeliveryUrlOperation
						r.summary = "Generate a signed imgproxy URL for delivery with requested params"
						r.operationID = "getDeliveryUrl"
						r.operationGroup = ""
						r.pathPattern = "/v1/images/{id}/url"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}

			}

		}
	}
	return r, false
}