# gen/typescript/api.schemas.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Image Service API
 * API for managing images
 * OpenAPI spec version: 1.0.30
 */
export type OwnerType = typeof OwnerType[keyof typeof OwnerType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OwnerType = {
  draft: 'draft',
  product: 'product',
  user: 'user',
} as const;

export type ImageRole = typeof ImageRole[keyof typeof ImageRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ImageRole = {
  main: 'main',
  gallery: 'gallery',
  other: 'other',
} as const;

export type ImageStatus = typeof ImageStatus[keyof typeof ImageStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ImageStatus = {
  uploaded: 'uploaded',
  ready: 'ready',
  deleted: 'deleted',
} as const;

export interface Variant {
  name?: string;
  /** S3 key of the variant */
  key?: string;
  width?: number;
  height?: number;
  mime?: string;
  size?: number;
}

export interface Image {
  id: string;
  ownerType: OwnerType;
  ownerId: string;
  role: ImageRole;
  /** S3 key of the original image */
  key: string;
  /**
   * Optional — direct URL if public or via proxy
   * @nullable
   */
  url?: string | null;
  alt: string;
  mime: string;
  size: number;
  /** @nullable */
  hash?: string | null;
  status: ImageStatus;
  variants?: Variant[];
  createdAt: string;
  modifiedAt: string;
}

export type PresignRequestContentType = typeof PresignRequestContentType[keyof typeof PresignRequestContentType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PresignRequestContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
  'image/avif': 'image/avif',
} as const;

export interface PresignRequest {
  ownerType: OwnerType;
  /**
   * Draft ID or owner ID
   * @minLength 1
   */
  ownerId: string;
  /** @minLength 1 */
  filename: string;
  contentType: PresignRequestContentType;
  /**
   * Size in bytes
   * @minimum 1
   * @maximum 10485760
   */
  size: number;
  role: ImageRole;
}

/**
 * Form fields to include with POST upload. Must be sent BEFORE the file field.
Includes: key, Content-Type, policy, x-amz-algorithm, x-amz-credential, x-amz-date, x-amz-signature

 */
export type PresignResponseFormData = {[key: string]: string};

export interface PresignResponse {
  /** URL to POST the file to (multipart/form-data) */
  uploadUrl: string;
  /** Signed JWT token for confirm endpoint (contains key, ownerType, ownerId, role, etc.) */
  uploadToken: string;
  /** TTL in seconds */
  expiresIn: number;
  /** Form fields to include with POST upload. Must be sent BEFORE the file field.
Includes: key, Content-Type, policy, x-amz-algorithm, x-amz-credential, x-amz-date, x-amz-signature
 */
  formData: PresignResponseFormData;
}

export interface ConfirmRequest {
  /**
   * Signed JWT token from presign response
   * @minLength 1
   */
  uploadToken: string;
  alt: string;
  role: ImageRole;
  /**
   * e.g., SHA256 hex for additional integrity check
   * @nullable
   */
  checksum?: string | null;
}

export interface PromoteRequest {
  /** @minLength 1 */
  draftId: string;
  /** @minLength 1 */
  productId: string;
  /** If true — copy to products/{productId}/... and remove old objects */
  move: boolean;
  /** If omitted — promote all images of the draft */
  images?: string[];
}

export type ProblemErrorsItem = {
  field?: string;
  message?: string;
};

/**
 * RFC7807 Problem Details
 */
export interface Problem {
  type: string;
  title: string;
  status: number;
  detail?: string;
  instance?: string;
  traceId?: string;
  errors?: ProblemErrorsItem[];
}

export type DeleteImageParams = {
/**
 * If true — delete objects in MinIO and remove DB record permanently
 */
hard?: boolean;
};

export type GetDeliveryUrlParams = {
/**
 * Name of a pre-generated variant (if exists)
 */
variant?: string;
/**
 * @minimum 1
 */
w?: number;
/**
 * @minimum 1
 */
h?: number;
fit?: GetDeliveryUrlFit;
format?: GetDeliveryUrlFormat;
/**
 * @minimum 1
 * @maximum 100
 */
quality?: number;
/**
 * @minimum 0.5
 * @maximum 3
 */
dpr?: number;
/**
 * If imgproxy is configured with signature expiration, an `expires` param will be included
 * @minimum 0
 */
ttlSeconds?: number;
};

export type GetDeliveryUrlFit = typeof GetDeliveryUrlFit[keyof typeof GetDeliveryUrlFit];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetDeliveryUrlFit = {
  cover: 'cover',
  contain: 'contain',
  fill: 'fill',
  inside: 'inside',
  outside: 'outside',
} as const;

export type GetDeliveryUrlFormat = typeof GetDeliveryUrlFormat[keyof typeof GetDeliveryUrlFormat];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetDeliveryUrlFormat = {
  original: 'original',
  webp: 'webp',
  avif: 'avif',
  jpeg: 'jpeg',
  png: 'png',
} as const;

export type GetDeliveryUrl200 = {
  url: string;
  expiresAt?: string;
};

export type PromoteImages200 = {
  promoted?: Image[];
};