# openapi/openapi.yml
openapi: 3.0.3
info:
  title: Image Service API
  version: 1.0.30
  description: API for managing images
paths:
  /v1/images/presign:
    post:
      summary: Create a POST policy for uploading an image to MinIO/S3 with size validation
      operationId: createPresign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresignRequest"
            examples:
              draftExample:
                value:
                  ownerType: draft
                  ownerId: dft_01JABCD12EF3G45H6JK7MN8PQ
                  filename: shoe.jpg
                  contentType: image/jpeg
                  size: 183456
                  role: main
      responses:
        "200":
          description: POST policy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignResponse"
              examples:
                ok:
                  value:
                    uploadUrl: "http://localhost:9000/products"
                    uploadToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    expiresIn: 900
                    formData:
                      key: "drafts/dft_01JABCD12EF3G45H6JK7MN8PQ/abc123.jpg"
                      Content-Type: "image/jpeg"
                      policy: "eyJleHBpcmF0aW9uIjoiMjAyNi0wMS..."
                      x-amz-algorithm: "AWS4-HMAC-SHA256"
                      x-amz-credential: "minioadmin/20260124/us-east-1/s3/aws4_request"
                      x-amz-date: "20260124T120000Z"
                      x-amz-signature: "a1b2c3d4e5f6..."
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  /v1/images/confirm:
    post:
      summary: Confirm upload (HEAD against S3), persist metadata
      operationId: confirmUpload
      # parameters:
      # - in: header
      #   name: Idempotency-Key
      #   schema:
      #     type: string
      #   required: false
      #   description: Idempotency key for safe retries.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
            examples:
              ok:
                value:
                  uploadToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  alt: "Black sneakers"
                  role: main
      responses:
        "201":
          description: Image document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Uploaded object not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: Conflict (e.g., image with the same key already confirmed)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  /v1/images/{id}:
    get:
      summary: Get image metadata
      operationId: getImage
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "404":
          description: Image Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

    delete:
      summary: Delete image (soft or hard, per policy)
      operationId: deleteImage
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: hard
          schema:
            type: boolean
            default: false
            description: If true — delete objects in MinIO and remove DB record permanently
      responses:
        "204":
          description: Deleted
        "404":
          description: Image Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  /v1/images/{id}/url:
    get:
      summary: Generate a signed imgproxy URL for delivery with requested params
      operationId: getDeliveryUrl
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: variant
          schema:
            type: string
            description: Name of a pre-generated variant (if exists)
        - in: query
          name: w
          schema:
            type: integer
            minimum: 1
        - in: query
          name: h
          schema:
            type: integer
            minimum: 1
        - in: query
          name: fit
          schema:
            type: string
            enum: [cover, contain, fill, inside, outside]
        - in: query
          name: format
          schema:
            type: string
            enum: [original, webp, avif, jpeg, png]
            default: original
        - in: query
          name: quality
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 85
        - in: query
          name: dpr
          schema:
            type: number
            minimum: 0.5
            maximum: 3
            default: 1
        - in: query
          name: ttlSeconds
          schema:
            type: integer
            minimum: 0
            description: If imgproxy is configured with signature expiration, an `expires` param will be included
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url:
                    type: string
                    format: uri
                  expiresAt:
                    type: string
                    format: date-time
        "404":
          description: Image Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  /v1/images/promote:
    post:
      summary: Promote images from a product draft to a final product
      operationId: promoteImages
      # parameters:
      #   - in: header
      #     name: Idempotency-Key
      #     schema:
      #       type: string
      #     required: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteRequest"
            examples:
              all:
                value:
                  fromDraft: dft_01JABCD12EF3G45H6JK7MN8PQ
                  toProduct: prd_01KXYZABCD1234
                  move: true
              subset:
                value:
                  fromDraft: dft_01JABCD12EF3G45H6JK7MN8PQ
                  toProduct: prd_01KXYZABCD1234
                  move: true
                  images:
                    - img_01MABCDEF...
                    - img_01MXYZ...
      responses:
        "200":
          description: Promotion completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  promoted:
                    type: array
                    items:
                      $ref: "#/components/schemas/Image"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Image Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: Conflict during promotion
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

components:
  schemas:
    OwnerType:
      type: string
      enum: [draft, product, user]
    ImageRole:
      type: string
      enum: [main, gallery, other]
    ImageStatus:
      type: string
      enum: [uploaded, ready, deleted]
    Variant:
      type: object
      properties:
        name:
          type: string
          example: thumb
        key:
          type: string
          description: S3 key of the variant
        width:
          type: integer
        height:
          type: integer
        mime:
          type: string
        size:
          type: integer
    Image:
      type: object
      required:
        [
          id,
          ownerType,
          ownerId,
          role,
          key,
          alt,
          mime,
          size,
          status,
          createdAt,
          modifiedAt,
        ]
      properties:
        id:
          type: string
          example: img_01MABCDEF...
        ownerType:
          $ref: "#/components/schemas/OwnerType"
        ownerId:
          type: string
        role:
          $ref: "#/components/schemas/ImageRole"
        key:
          type: string
          description: S3 key of the original image
        url:
          type: string
          nullable: true
          description: Optional — direct URL if public or via proxy
        alt:
          type: string
        mime:
          type: string
        size:
          type: integer
        hash:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ImageStatus"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/Variant"
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
    PresignRequest:
      type: object
      required: [ownerType, ownerId, filename, contentType, size, role]
      properties:
        ownerType:
          $ref: "#/components/schemas/OwnerType"
        ownerId:
          type: string
          minLength: 1
          description: Draft ID or owner ID
        filename:
          type: string
          minLength: 1
          example: shoe.jpg
        contentType:
          type: string
          enum: [image/jpeg, image/png, image/webp, image/avif]
          example: image/jpeg
        size:
          type: integer
          minimum: 1
          maximum: 10485760 # 10 MiB default limit; adjust as needed
          description: Size in bytes
        role:
          $ref: "#/components/schemas/ImageRole"
    PresignResponse:
      type: object
      required: [uploadUrl, uploadToken, expiresIn, formData]
      properties:
        uploadUrl:
          type: string
          format: uri
          description: URL to POST the file to (multipart/form-data)
        uploadToken:
          type: string
          description: Signed JWT token for confirm endpoint (contains key, ownerType, ownerId, role, etc.)
        expiresIn:
          type: integer
          description: TTL in seconds
        formData:
          type: object
          additionalProperties:
            type: string
          description: |
            Form fields to include with POST upload. Must be sent BEFORE the file field.
            Includes: key, Content-Type, policy, x-amz-algorithm, x-amz-credential, x-amz-date, x-amz-signature
    ConfirmRequest:
      type: object
      required: [uploadToken, alt, role]
      properties:
        uploadToken:
          type: string
          minLength: 1
          description: Signed JWT token from presign response
        alt:
          type: string
        role:
          $ref: "#/components/schemas/ImageRole"
        checksum:
          type: string
          nullable: true
          description: e.g., SHA256 hex for additional integrity check
    PromoteRequest:
      type: object
      required: [draftId, productId, move]
      properties:
        draftId:
          type: string
          minLength: 1
        productId:
          type: string
          minLength: 1
        move:
          type: boolean
          description: If true — copy to products/{productId}/... and remove old objects
        images:
          type: array
          description: If omitted — promote all images of the draft
          items:
            type: string
    Problem:
      type: object
      description: RFC7807 Problem Details
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }